l2-ingestion/requirements.txt
# L2 Parser
pyyaml>=6.0
PyMuPDF>=1.24.0
pdfminer.six>=20221105
inotify_simple>=1.3.5
httpx>=0.27.0

# L2-07 OTEL + FastAPI
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
opentelemetry-api>=1.24.0
fastapi>=0.111.0
uvicorn>=0.29.0

# L2-03 Embedder
qdrant-client>=1.9.0
